requirements.txt
django==6.0
    
django-environ==0.12.0

djangorestframework-simplejwt==5.5.1

django-tailwind==4.4.2

Pillow

gunicorn

whitenoise

